Rename false spike to wrong spike in configs
The word "wrong" fits better for enemy spikes than "false".
To me it just sounds better. And to quote the native player Ewan:

> they’re equally correct imo
> but when talking about teams i’d tend towards wrong
> bcs there’s no true team, just the one you’re on

Possible alternatives are:
- opponent
- enemey

But wrong sounds the most obvious in all the cases. Otherwise
it can be confused with for example spiking an enemy. Not with the
spike that belongs to the enemy. Also what does enemy spike mean?
The spike where you put the enemys? The spike where the enemys put you?
The word "wrong" is clear in these cases.

This came up in ddnet-insta where the same config is being implemented.
And I would like to share as much user facing variable names as possible
to make the life of the users easier and more consistent.

https://github.com/ddnet-insta/ddnet-insta/pull/327#discussion_r1994325106